Baseline- forwards evdev to uniput with matrix bypass.
This is a rough baseline capability that takes input events and
forwards them to uinput. There's also a basic virtual matrix for
the keypad that intercepts those key events.

Next steps look a lot like the python prototyping did:
- Add a basic layers capability.
- Add advanced key codes and grow the driver to accomodate.

The project also needs to be broken into a library and a binary.
That splitting process needs to also include documenting what's
been written.