Implemented string -> key code logic.
This was a significant change that impacted many lines. The feature
was the logic to convert from a string to a key code. For example:

SPACECADET(KC_A, KC_LEFTSHIFT)

This feature is necessary for creating run-time keyboard layouts
(instead of only supporting hard coded, QMK style compile-time
layouts).

This feature involved writing a simple parser that takes the string
and converts it into an expression tree. Every possible key code
then tries to convert the expression tree into something it
understands.

This feature forced me to reconsider whether or not keys should be
an enum or a boxed trait. I stuck with boxed trait for a good
reason that I can't recall (the experiment wound up in a dark
corner).

This feature is incomplete - here are the remaining TODOs:

- It's missing tests for both the tokenizer and the tree -> code
conversion.

- The list of possible keys is hard coded, which effectively
prevents users from extending the library with new keys.